Refactored to bring WOAH on front.
