+++ ACL2020/data/CC_supplementary.pdf
